Remove the intermediate .o file
It turns out that the linker will exclude objects from an archive if it detects
none of the functions in that object are used. Because I was making a single
big object before adding it to the archive it would import all the object code
into the binary. By splitting this I reduced my test binary from 133800 bytes
to 68696 bytes because I was only using mat4 functions.